refactor: relocate catalog operations into setup and common
