+++ src/pipeline/bronze/ingest_jepx.py
@@ -9,7 +9,7 @@ import polars as pl
 
 sys.path.append(str(Path(__file__).resolve().parent.parent.parent))
 
-from catalog.manage_iceberg import get_catalog
+from common.iceberg import get_catalog
 from common.jepx_common import (
     resolve_spot_summary_file_name,
     resolve_spot_summary_object_key,
