+++ src/setup/drop_table.py
