+++ src/setup/provision.py
@@ -3,7 +3,8 @@ from pathlib import Path
 
 from pyiceberg.catalog import load_catalog
 from pyiceberg.exceptions import NamespaceAlreadyExistsError, NoSuchTableError
-from schema_builder import build_table_schema
+
+from common.schema_builder import build_table_schema
 
 # ログの中央集権設定(ここで出力形式を固定する)
 logging.basicConfig(
